Change `Record` to dataclass
Need methods in `Record` therefore it cannot be NamedTuple. Additionally, two
more fields are added to `Record`: "origin" and "currency".

Implementation-wise there is no need to compare instances and `order=True` in
`Record` because this is more expensive. Sorting can be made on a lower level
with sorting keys and `bisect`.